backend/requirements.txt
numpy
pandas
python-dotenv==0.19.1
psycopg2-binary
flask
gunicorn
requests
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn
datasets
apache-beam
apache-airflow[postgres,aws]
apache-airflow-providers-amazon
dvc
dvc[s3]
boto3
ydata-profiling
awscli